Feature/database

+++ src/Database.py
@@ -1,29 +1,70 @@
 import sqlite3 as sqlite
+import datetime
 import os.path
 
+TIME_FORMAT = '%Y-%m-%d %H:%M:%S'
 # SQL
-SQL_GET_LIST = "SELECT * FROM task WHERE status = (?)"
-SQL_DELETE_TODO = "DELETE FROM task WHERE id=(?)"
+## CREATE
+SQL_CREATE_TABLE = "CREATE TABLE IF NOT EXISTS {}(id INTEGER PRIMARY KEY, status TEXT, content TEXT, startTime TEXT, endTime TEXT)"
+SQL_INSERT_TODO = "INSERT INTO {}(status, content) values (?,?)"
+SQL_GET_LIST = "SELECT * FROM {} WHERE status = (?)"
+SQL_UPDATE_TODO = "UPDATE {} SET status = (?) WHERE id=(?)"
+SQL_UPDATE_START_TIME = "UPDATE {} SET startTime = (?) WHERE id=(?)"
+SQL_UPDATE_END_TIME = "UPDATE {} SET endTime = (?) WHERE id=(?)"
 
-SQL_UPDATE_TODO = "UPDATE task SET status = (?) WHERE id=(?)"
-SQL_UPDATE_START_TIME = "UPDATE task SET startTime = (?) WHERE id=(?)"
-SQL_UPDATE_END_TIME = "UPDATE task SET endTime = (?) WHERE tid=(?)"
-
-SQL_INSERT_TODO = "INSERT INTO task(status, description) values (?,?)"
-SQL_CREATE_TABLE = "CREATE TABLE IF NOT EXISTS task(id INTEGER PRIMARY KEY, status TEXT, description TEXT, startTime TEXT, endTime TEXT)"
 SQL_USED_TIME = "SELECT strftime('%s','now', 'localtime') - strftime('%s',(?));"
-
+SQL_DELETE_TODO = "DELETE FROM {} WHERE id=(?)"
 
 class Database:
     conn = None
+    databaseName = None
+    tableName = None
 
     def createDatabase(self, databaseName):
         self.conn = sqlite.connect(databaseName)
+        self.databaseName = databaseName
 
     def removeDatabase(self, databaseName):
-        os.remove(databaseName)
+        if os.path.exists(databaseName):
+            os.remove(databaseName)
+            self.databaseName = None
 
     def createTable(self, tableName):
+        self.tableName = tableName
+        SQL = SQL_CREATE_TABLE.format(tableName)
         with self.conn:
-            cur = conn.cursor()
-            cur.execute(tableName)
+            cur = self.conn.cursor()
+            cur.execute(SQL)
+    
+    def insertTodo(self, status, content):
+        SQL = SQL_INSERT_TODO.format(self.tableName)
+        with self.conn:
+            cur = self.conn.cursor()
+            cur.execute(SQL, [status, content])
+
+    def getTodoList(self, status):
+        SQL = SQL_GET_LIST.format(self.tableName)
+        with self.conn:
+            cur = self.conn.cursor()
+            cur.execute(SQL, [status])
+            return cur.fetchall()
+
+    def updateStatus(self, index, status):
+        SQL = SQL_UPDATE_TODO.format(self.tableName)
+        with self.conn:
+            cur = self.conn.cursor()
+            cur.execute(SQL, [status, index])
+
+    def updateStartTime(self, index):
+        SQL = SQL_UPDATE_START_TIME.format(self.tableName)
+        time = datetime.datetime.now().strftime(TIME_FORMAT)
+        with self.conn:
+            cur = self.conn.cursor()
+            cur.execute(SQL, [time, index])
+
+    def updateEndTime(self, index):
+        SQL = SQL_UPDATE_END_TIME.format(self.tableName)
+        time = datetime.datetime.now().strftime(TIME_FORMAT)
+        with self.conn:
+            cur = self.conn.cursor()
+            cur.execute(SQL, [time, index])
